move 3d parseColorScale to gl_format_color
